refactor(server): split resolvers into per-entity files and migrate to Drizzle 1.0 query API
Break the monolithic resolvers.ts (1488 lines) into focused files under
schema/resolvers/ (profile, activity-types, todos, habits, time-blocks,
stats, schedule, auth). Fix the DrizzleError caused by passing SQL
expression objects to db.query.* where clauses — migrate all relational
queries to the new Drizzle 1.0 object filter syntax ({ userId },
{ completedAt: { isNull: true } }, { id: { in: [...] } }, OR/AND) and
object orderBy ({ priority: 'desc' }). Use relation filter in
scheduler-writeback to scope habitCompletions via habit.userId directly,
eliminating a redundant inner query.
