+++ package.json
@@ -15,8 +15,8 @@
     "lint": "biome check .",
     "lint:fix": "biome check --apply .",
     "db:generate": "npm run db:generate -w @auto-cal/db",
-    "db:migrate": "npm run migrate -w @auto-cal/db",
-    "db:seed": "npm run seed -w @auto-cal/db",
+    "db:migrate": "npm run db:migrate -w @auto-cal/db",
+    "db:seed": "npm run db:seed -w @auto-cal/db",
     "db:studio": "npm run db:studio -w @auto-cal/db",
     "codegen": "graphql-codegen --config codegen.ts",
     "codegen:server": "graphql-codegen-esm --config codegen.server.ts",
